R2 management team requirement by cosmos

+++ app/Http/Controllers/BuildingController.php
@@ -6,6 +6,7 @@ use App\Building;
 use Illuminate\Support\Collection;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
+use Illuminate\Support\Facades\DB;
 
 use App\Responser\NestedRelationResponser;
 use App\Responser\FormDataResponser;
@@ -31,10 +32,16 @@ class BuildingController extends Controller
         // in case of might need to display nested resources while listing
         // ex:  $responseData->relations(['rooms'])
 
+        $selectStr    = DB::raw(join(', ', array_merge(['buildings.*'], Building::extraInfoColumns())));
         $responseData = new NestedRelationResponser();
         $responseData
-            ->index('buildings',
-                $this->limitRecords(Building::with($request->withNested))
+            ->index(
+                'buildings',
+                $this->limitRecords(
+                    Building::withExtraInfo()
+                        ->select(DB::raw($selectStr))
+                        ->with($request->withNested)
+                )
             )
             ->relations($request->withNested);
 
@@ -49,11 +56,15 @@ class BuildingController extends Controller
     public function create(Request $request)
     {
         $responseData = new FormDataResponser();
-        $data = $responseData
+        $responseData = $responseData
             ->create(Building::class, 'buildings.store')
             ->get();
 
-        return view('buildings.form', $data);
+        if ($request->old()) {
+            $responseData['data'] = array_merge($responseData['data'], $request->old());
+        }
+
+        return view('buildings.form', $responseData);
     }
 
     /**
@@ -65,57 +76,61 @@ class BuildingController extends Controller
     public function store(Request $request)
     {
         $validatedData = $request->validate([
-            'title' => 'required|max:255',
+            'title'         => 'required|max:255',
             'building_code' => 'required|max:255',
-            'group' => 'required|max:255',
-            'city' => [
+            'group'         => 'max:255',
+            'city'          => [
                 'required',
                 Rule::in(array_keys(config('enums.cities')))
             ],
             'district' => [
                 'bail',
                 'required_with:city',
-                Rule::in(config('enums.cities.' . $request->city))
+                Rule::in(config('enums.cities.'.$request->city))
             ],
             'address' => 'required|max:255',
-            'tax_number' => 'required|max:255',
-            'building_type' => 'required|max:255',
-            'floor' => 'required|integer|digits_between:1,11',
-            'legal_usage' => 'required|max:255',
-            'has_elevator' => 'required|boolean',
-            'security_guard' => 'required|max:255',
-            'management_count' => 'required|max:255',
-            'first_floor_door_opening' => 'required|max:255',
-            'public_area_door_opening' => 'required|max:255',
-            'room_door_opening' => 'required|max:255',
-            'main_ammeter_location' => 'required|max:255',
-            'ammeter_serial_number_1' => 'required|max:255',
-            'shared_electricity' => 'required|max:255',
-            'taiwan_electricity_payment_method' => 'required|max:255',
-            'electricity_payment_method' => [
-                'required',
-                Rule::in(
-                    config('enums.buildings.electricity_payment_method')
-                )
+
+            'is_squatter'     => 'boolean',
+            'squatter_status' => 'max:255',
+            'decoration_needed' => 'required|boolean',
+            'decoration_price' => 'integer',
+
+            'tax_number'                        => 'required|max:255',
+            'building_type'                     => 'max:255',
+            'floor'                             => 'required|integer|digits_between:1,11',
+            'legal_usage'                       => 'max:255',
+            'land_use'                          => 'max:255',
+            'has_elevator'                      => 'boolean',
+            'security_guard'                    => 'max:255',
+            'management_count'                  => 'required|max:255',
+            'first_floor_door_opening'          => 'max:255',
+            'public_area_door_opening'          => 'max:255',
+            'room_door_opening'                 => 'max:255',
+            'main_ammeter_location'             => 'required|max:255',
+            'ammeter_serial_number_1'           => 'required|max:255',
+            'shared_electricity'                => 'max:255',
+            'taiwan_electricity_payment_method' => 'max:255',
+            'electricity_payment_method'        => [
+                Rule::in(config('enums.buildings.electricity_payment_method'))
             ],
-            'private_ammeter_location' => 'required|max:255',
-            'water_meter_location' => 'required|max:255',
-            'water_meter_serial_number' => 'required|max:255',
-            'water_payment_method' => 'required|max:255',
-            'water_meter_reading_date' => 'required|date',
-            'gas_meter_location' => 'required|max:255',
-            'garbage_collection_location' => 'required|max:255',
-            'garbage_collection_time' => 'required|max:255',
-            'management_fee_payment_method' => 'required|max:255',
-            'management_fee_contact' => 'required|max:255',
-            'management_fee_contact_phone' => 'required|max:255',
-            'distribution_method' => 'required|max:255',
-            'administrative_number' => 'required|max:255',
-            'accounting_group' => 'required|max:255',
-            'rental_receipt' => 'required|max:255',
-            'commissioner_id' => 'nullable|exists:users,id',
-            'administrator_id' => 'nullable|exists:users,id',
-            'comment' => 'required|max:255'
+            'private_ammeter_location'      => 'max:255',
+            'water_meter_location'          => 'required|max:255',
+            'water_meter_serial_number'     => 'required|max:255',
+            'water_payment_method'          => 'required|max:255',
+            'water_meter_reading_date'      => 'required|date',
+            'gas_meter_location'            => 'required|max:255',
+            'garbage_collection_location'   => 'max:255',
+            'garbage_collection_time'       => 'max:255',
+            'management_fee_payment_method' => 'max:255',
+            'management_fee_contact'        => 'required|max:255',
+            'management_fee_contact_phone'  => 'required|max:255',
+            'distribution_method'           => 'max:255',
+            'administrative_number'         => 'required|max:255',
+            'accounting_group'              => 'max:255',
+            'rental_receipt'                => 'max:255',
+            'commissioner_id'               => 'exists:users,id',
+            'administrator_id'              => 'exists:users,id',
+            'comment'                       => 'max:255'
         ]);
 
         $newBuilding = BuildingService::create($validatedData);
@@ -146,13 +161,16 @@ class BuildingController extends Controller
      * @param  \App\Building  $building
      * @return \Illuminate\Http\Response
      */
-    public function edit(Building $building)
+    public function edit(Request $request, Building $building)
     {
         $responseData = new FormDataResponser();
-        return view(
-            'buildings.form',
-            $responseData->edit($building, 'buildings.update')->get()
-        );
+        $responseData = $responseData->edit($building, 'buildings.update')->get();
+
+        if ($request->old()) {
+            $responseData['data'] = array_merge($responseData['data'], $request->old());
+        }
+
+        return view('buildings.form', $responseData);
     }
 
     /**
@@ -165,57 +183,61 @@ class BuildingController extends Controller
     public function update(Request $request, Building $building)
     {
         $validatedData = $request->validate([
-            'title' => 'required|max:255',
+            'title'         => 'required|max:255',
             'building_code' => 'required|max:255',
-            'group' => 'required|max:255',
-            'city' => [
+            'group'         => 'max:255',
+            'city'          => [
                 'required',
                 Rule::in(array_keys(config('enums.cities')))
             ],
             'district' => [
                 'bail',
                 'required_with:city',
-                Rule::in(config('enums.cities.' . $request->city))
+                Rule::in(config('enums.cities.'.$request->city))
             ],
             'address' => 'required|max:255',
-            'tax_number' => 'required|max:255',
-            'building_type' => 'required|max:255',
-            'floor' => 'required|integer|digits_between:1,11',
-            'legal_usage' => 'required|max:255',
-            'has_elevator' => 'required|boolean',
-            'security_guard' => 'required|max:255',
-            'management_count' => 'required|max:255',
-            'first_floor_door_opening' => 'required|max:255',
-            'public_area_door_opening' => 'required|max:255',
-            'room_door_opening' => 'required|max:255',
-            'main_ammeter_location' => 'required|max:255',
-            'ammeter_serial_number_1' => 'required|max:255',
-            'shared_electricity' => 'required|max:255',
-            'taiwan_electricity_payment_method' => 'required|max:255',
-            'electricity_payment_method' => [
-                'required',
-                Rule::in(
-                    config('enums.buildings.electricity_payment_method')
-                )
+
+            'is_squatter'     => 'boolean',
+            'squatter_status' => 'max:255',
+            'decoration_needed' => 'required|boolean',
+            'decoration_price' => 'integer',
+
+            'tax_number'                        => 'required|max:255',
+            'building_type'                     => 'max:255',
+            'floor'                             => 'required|integer|digits_between:1,11',
+            'legal_usage'                       => 'max:255',
+            'land_use'                          => 'max:255',
+            'has_elevator'                      => 'boolean',
+            'security_guard'                    => 'max:255',
+            'management_count'                  => 'required|max:255',
+            'first_floor_door_opening'          => 'max:255',
+            'public_area_door_opening'          => 'max:255',
+            'room_door_opening'                 => 'max:255',
+            'main_ammeter_location'             => 'required|max:255',
+            'ammeter_serial_number_1'           => 'required|max:255',
+            'shared_electricity'                => 'max:255',
+            'taiwan_electricity_payment_method' => 'max:255',
+            'electricity_payment_method'        => [
+                Rule::in(config('enums.buildings.electricity_payment_method'))
             ],
-            'private_ammeter_location' => 'required|max:255',
-            'water_meter_location' => 'required|max:255',
-            'water_meter_serial_number' => 'required|max:255',
-            'water_payment_method' => 'required|max:255',
-            'water_meter_reading_date' => 'required|date',
-            'gas_meter_location' => 'required|max:255',
-            'garbage_collection_location' => 'required|max:255',
-            'garbage_collection_time' => 'required|max:255',
-            'management_fee_payment_method' => 'required|max:255',
-            'management_fee_contact' => 'required|max:255',
-            'management_fee_contact_phone' => 'required|max:255',
-            'distribution_method' => 'required|max:255',
-            'administrative_number' => 'required|max:255',
-            'accounting_group' => 'required|max:255',
-            'rental_receipt' => 'required|max:255',
-            'commissioner_id' => 'nullable|exists:users,id',
-            'administrator_id' => 'nullable|exists:users,id',
-            'comment' => 'required|max:255'
+            'private_ammeter_location'      => 'max:255',
+            'water_meter_location'          => 'required|max:255',
+            'water_meter_serial_number'     => 'required|max:255',
+            'water_payment_method'          => 'required|max:255',
+            'water_meter_reading_date'      => 'required|date',
+            'gas_meter_location'            => 'required|max:255',
+            'garbage_collection_location'   => 'max:255',
+            'garbage_collection_time'       => 'max:255',
+            'management_fee_payment_method' => 'max:255',
+            'management_fee_contact'        => 'required|max:255',
+            'management_fee_contact_phone'  => 'required|max:255',
+            'distribution_method'           => 'max:255',
+            'administrative_number'         => 'required|max:255',
+            'accounting_group'              => 'max:255',
+            'rental_receipt'                => 'max:255',
+            'commissioner_id'               => 'exists:users,id',
+            'administrator_id'              => 'exists:users,id',
+            'comment'                       => 'max:255'
         ]);
 
         $building->update($validatedData);
@@ -229,19 +251,33 @@ class BuildingController extends Controller
      * @param  \App\Building  $building
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Building $building)
+    public function destroy(Request $request, Building $building)
     {
+        $user = auth()->user();
+
+        $permission = $user->getAllPermissions()->first(function ($permission) {
+            return $permission->name = 'delete building';
+        });
+
+        if (! $permission) {
+            return response()->json([
+                'errors' => ['permission denied']
+            ], 403);
+        }
+
         $building->delete();
+
         return response()->json(true);
     }
 
-    public function electricityPaymentReport(Building $building, int $year, int $month) {
+    public function electricityPaymentReport(Building $building, int $year, int $month)
+    {
         $reportRows = $this->buildElectricityPaymentReportData($building, $year, $month);
 
         return view('buildings.electricity_payment_report', [
             'reportRows' => $reportRows,
-            'year' => $year,
-            'month' => $month,
+            'year'       => $year,
+            'month'      => $month,
         ]);
     }
 
@@ -254,9 +290,12 @@ class BuildingController extends Controller
      *
      * @return array
      */
-    private function buildElectricityPaymentReportData(Building $building, int $year, int $month): Collection {
+    private function buildElectricityPaymentReportData(Building $building, int $year, int $month): Collection
+    {
         $rooms = $building->rooms()->get();
 
-        return $rooms->map(function ($room) use ($year, $month) { return $room->buildElectricityPaymentData($year, $month); });
+        return $rooms->map(function ($room) use ($year, $month) {
+            return $room->buildElectricityPaymentData($year, $month);
+        });
     }
 }
